requirements.txt
quart
quart-auth
peewee
pillow-simd
python-magic
aiofiles
hypercorn
python-dateutil